requirements.txt
streamlit==0.58.0
numpy==1.16.4
pandas==1.0.3
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz#egg=en_core_web_sm
matplotlib==3.2.1
Pillow==7.0.0
wordcloud==1.6.0
gTTS==2.1.1
sumy==0.8.1
regex==2020.1.8
beautifulsoup4==4.6.0
urllib3==1.22
requests==2.18.4
requests-file==1.4.3


